add result.hpp
